+++ tests/live_tests/scenarios/global/base.py
@@ -34,7 +34,7 @@ class BasicScenario(LiveScenario):
 
     AS_SET = {
         "AS-AS1": [1],
-        "AS-AS1_CUSTOMERS": [101, 103],
+        "AS-AS1_CUSTOMERS": [101, 103, 104],
         "AS-AS2": [2],
         "AS-AS2_CUSTOMERS": [101, 103],
     }
@@ -233,6 +233,23 @@ class BasicScenario(LiveScenario):
                            self.AS1_1, as_path="1 101", next_hop=self.AS1_1)
         self.receive_route(self.rs, prefix,
                            self.AS2, as_path="2 101", next_hop=self.AS2)
+
+    def test_030_good_prefixes_because_of_arin_db_1(self):
+        """{}: good prefixes because of use_arin_bulk_whois_data"""
+
+        prefix = self.DATA["AS104_arin_1"]
+        self.receive_route(self.rs, prefix,
+                           self.AS1_1, as_path="1 101 104",
+                           next_hop=self.AS1_1)
+
+        # AS101 peers with AS2 too but that route should be reject
+        # because only AS1 has AS104 in its AS-SET
+        self.receive_route(self.rs, prefix,
+                           self.AS2, as_path="2 101 104", next_hop=self.AS2,
+                           filtered=True, reject_reason=9)
+
+        self.receive_route(self.AS4, prefix, self.rs, as_path="1 101 104",
+                           next_hop=self.AS1_1)
 
     def test_030_good_prefixes_received_by_rs_nonclient_nexthop(self):
         """{}: good prefixes received by rs: non-client NEXT_HOP"""
@@ -919,7 +936,7 @@ class BasicScenario(LiveScenario):
         """{}: control communities, RTT, prepend 3x <= 5 ms, 2x <= 20 ms, 1x any"""
         pref = self.DATA["AS4_rtt_9"]
         self.receive_route(self.rs, pref, self.AS4,
-                           std_comms=["64536:5", "64535:20", "999:65501"])
+                           std_comms=["64536:5", "64535:20", "65521:65521"])
         for inst in [self.AS1_1, self.AS1_2]:
             self.receive_route(inst, pref, self.rs, as_path="4 4 4 4",
                                std_comms=[], lrg_comms=[], ext_comms=[])
@@ -990,7 +1007,7 @@ class BasicScenario(LiveScenario):
         # also verifies that these prefixes are received twice by AS3)
         for pref in ("AS101_good1", "AS101_bad_std_comm", "AS101_bad_lrg_comm",
                      "AS101_other_s_comm", "AS101_other_l_comm",
-                     "AS101_bad_good_comms"):
+                     "AS101_bad_good_comms", "AS104_arin_1"):
             self.receive_route(self.AS3, self.DATA[pref], self.rs)
 
         if not isinstance(self.rs, OpenBGPDInstance):