+++ CHANGES.rst
@@ -3,6 +3,22 @@ Change log
 
 .. note:: **Upgrade notes**: after upgrading, run the ``arouteserver setup-templates`` command to sync the local templates with those distributed with the new version. More details on the `Upgrading <https://arouteserver.readthedocs.io/en/latest/INSTALLATION.html#upgrading>`__ section of the documentation.
 
+next release
+------------
+
+- Improvement: OpenBGPD, more flexibility for inbound communities values.
+
+  This allows to use inbound 'peer_as' communities which overlap with other inbound communities whose last *part* is a private ASN.
+
+- New feature: use ARIN Whois database dump to authorize routes.
+
+  This feature allows to accept those routes whose origin ASN is authorized by a client AS-SET, whose prefix has not a corresponding route object but is covered by an ARIN Whois record for the same origin ASN.
+
+- Improvement: extend the use of *RPKI ROAs as route objects* and *ARIN Whois database dump* to ``tag_as_set``-only mode.
+
+  Before of this, the *RPKI ROAs as route objects* and *ARIN Whois DB dump* features were used only when origin AS and prefix enforcing was set.
+  Starting with this release they are used even when enforcing is not configured and only the ``tag_as_set`` mode is used.
+
 v0.15.0
 -------